app/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
psycopg2-binary>=2.9
django-cors-headers>=4.4
python-decouple>=3.8
xhtml2pdf